Isolate CoverImage into a separate file
