+++ app/javascript/article-form/components/CoverImage.jsx
@@ -1,0 +1,12 @@
+import { h } from 'preact';
+import { CoverImagePropTypes } from '../../src/components/common-prop-types';
+
+const CoverImage = ({ className, imageSrc, imageAlt }) => (
+  <div className={className}>
+    <img src={imageSrc} alt={imageAlt} />
+  </div>
+);
+
+CoverImage.propTypes = CoverImagePropTypes;
+
+export default CoverImage;
