+++ app/javascript/src/components/common-prop-types/cover-image-prop-types.js
@@ -1,0 +1,7 @@
+import PropTypes from 'prop-types';
+
+export const CoverImagePropTypes = PropTypes.shape({
+  className: PropTypes.string.isRequired,
+  imageSrc: PropTypes.string.isRequired,
+  imageAlt: PropTypes.string.isRequired,
+});
