+++ app/javascript/src/components/common-prop-types/index.js
@@ -3,3 +3,4 @@ export * from './default-children-prop-types';
 export * from './organization-prop-type';
 export * from './tag-prop-types';
 export * from './article-prop-types';
+export * from './cover-image-prop-types';
